+++ jest.config.base.js
@@ -1,4 +1,6 @@
 const baseConfig = require('./jest.config.base')
+const fs = require('fs')
+const config = JSON.parse(fs.readFileSync(`${__dirname}/.swcrc`, 'utf-8'))
 
 module.exports = {
   ...baseConfig,
@@ -8,7 +10,7 @@ module.exports = {
     '**/?(*.)+(spec|test).+(ts|tsx|js)',
   ],
   transform: {
-    '^.+\\.(ts|tsx)$': 'ts-jest',
+    '^.+\\.(t|j)sx?$': ['@swc/jest', { ...config, swcrc: false, exclude: [] }],
   },
   moduleNameMapper: {
     '^@/(.*)': '<rootDir>/src/$1',