+++ packages/fetcher/src/helpers/request.test.ts
@@ -1,7 +1,7 @@
-import { FetcherAbortManager } from '../abort-manager'
+import type { FetcherAbortManager } from '../abort-manager'
 import { DEFAULT_CONFIG } from '../const'
-import { HTTP_METHODS } from '../enums'
-import { FetcherConfig, FetcherRequestConfig } from '../types'
+import type { HTTP_METHODS } from '../enums'
+import type { FetcherConfig, FetcherRequestConfig } from '../types'
 import { buildRequestConfig, buildRequestURL } from './request'
 
 const CFG: FetcherConfig = {