+++ packages/w3p/src/provider.ts
@@ -1,8 +1,8 @@
-import { PROVIDERS } from '@/enums'
+import type { PROVIDERS } from '@/enums'
 import { errors } from '@/errors'
 import { ProviderDetector } from '@/provider-detector'
 
-import {
+import type {
   Chain,
   ChainId,
   IProvider,