+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- `@distributedlab/jac` - Re-export `@distributedlab/fetcher` `HTTP_METHODS, HTTP_STATUS_CODES` enums
 - `@distributedlab/w3p` - wrapper for web3 providers
   - `EVN based`:
     - `MetamaskProvider`
@@ -16,9 +17,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `Near based`:
     - `NearProvider`
 
+### Changed
+- `all` - Migrate from `tsc` to `swc` to build packages
+- `all` - Migrate from `ts-jest` to `swc-jest` to run tests
+
 ### Fixed
 - `root` - CI prebuild packages to resolve `@distributedlab/fetcher` dependency in the `@distributedlab/jac` package
 - `root` - Prevent typedoc from detecting every index file as a module
+- `all` - Replace type imports with `import type` syntax according to the [SWC tsc migration guide]
+- `all` - Set "browser": to "./dist/esm/index.js" to correctly resolve the package by the Webpack
+
+[SWC tsc migration guide]: https://swc.rs/docs/migrating-from-tsc
 
 ## [0.2.0-rc.1] - 2023-04-17
 ### Added